Fixes unlimited pages in pagination
This commit adds a file called functions.py which will hold any
function needed to tweak the theme. In this file is defined a funcition
called REPAGINATE, which receives as arguments a list (holding all
pages), a page number (current page), and a number (which defines the
max number of elements in the pagination). This function returns a new
pagination list with limited elements. The number of pages available in
the pagination can be adjusted by the MAX_PAGES_IN_PAGINATION variable
in the pelicanconf.py file.